Fix free

+++ include/utility/heap.h
@@ -79,14 +79,14 @@ public:
     }
 
     static void typed_free(void * ptr) {
-        int * addr = reinterpret_cast<int *>(ptr);
+        long * addr = reinterpret_cast<long *>(ptr);
         unsigned int bytes = *--addr;
         Heap * heap = reinterpret_cast<Heap *>(*--addr);
         heap->free(addr, bytes);
     }
 
     static void untyped_free(Heap * heap, void * ptr) {
-        int * addr = reinterpret_cast<int *>(ptr);
+        long * addr = reinterpret_cast<long *>(ptr);
         unsigned int bytes = *--addr;
         heap->free(addr, bytes);
     }
